Add -o offset, -a search options to zdb for use with the -l flag
The -l option is modified to accept an offset parameter which tells it to
read ZFS labels from a offset bytes into the device if possible.

The -o option modifies -l to read the ZFS device label from an offset into the
given device. The actual labels are read from their normal position presuming
the offset is the start of the device.

The -a option causes zdb to try and read ZFS device labels starting from every
byte past the offset, and stops once a valid label is found.

This provides the a byte-offset boundary for creating a containing partition
to allow for successful pool imports.